Split the gamemode trackers into multiple smaller components
